[avf] Add e2e test to check AVF key provisioning
This cl adds an e2e test that interacts with the real RKP
server to ensure that the AVF keys are successfully
provisioned on the device.

The test is forked from RkpdAppIntegrationTests.

The test target will be added to avf-presubmit later to ensure
that it runs on a real arm64 device in pre-submit.

Bug: 325610326
Test: atest AvfRkpdAppIntegrationTests
Change-Id: Idb3a54bce46b3dde82df6a1bf11e110e1f033cec